+++ Gamitour/WebContent/Procesos/MostrarParadas.jsp
@@ -62,17 +62,16 @@
 						<td>${parada.gastronomia}</td>
 						<td>${parada.imagen}</td>
 						<td>${parada.itinerario.getNombre()}</td>
-						<td><a id="eliminar"
+						<td><a class="eliminar" onclick="return confirm('¿Está seguro de que quiere continuar?')"
 							href="EliminarParadas.do?id=${parada.idparada}">Eliminar</a></td>
-						<td><a id="actualizar"
+						<td><a class="actualizar"
 							href="Procesos/Modificar/ModificarParadas.jsp?id=${parada.idparada}">Modificar</a></td>
 					</tr>
 				</c:forEach>
 			</table>
-			<a id="insertar" href="Procesos/Insertar/InsertarParadas.jsp">Insertar una
+			<a class="insertar" href="Procesos/Insertar/InsertarParadas.jsp">Insertar una
 				parada nueva</a>
 		</div>
-
 	</div>
 
 	<div id="footer"></div>
